Fix object field types (#12)

+++ TiledSharp/src/ObjectGroup.cs
@@ -50,26 +50,26 @@ namespace TiledSharp
             public string Name {get; private set;}
             public TmxObjectType ObjectType {get; private set;}
             public string Type {get; private set;}
-            public int X {get; private set;}
-            public int Y {get; private set;}
-            public int Width {get; private set;}
-            public int Height {get; private set;}
+            public double X {get; private set;}
+            public double Y {get; private set;}
+            public double Width {get; private set;}
+            public double Height {get; private set;}
             public double Rotation {get; private set;}
             public TmxLayerTile Tile {get; private set;}
             public bool Visible {get; private set;}
 
-            public List<Tuple<int,int>> Points {get; private set;}
+            public List<Tuple<double, double>> Points {get; private set;}
             public PropertyDict Properties {get; private set;}
 
             public TmxObject(XElement xObject)
             {
                 Name = (string)xObject.Attribute("name") ?? "";
                 Type = (string)xObject.Attribute("type");
-                X = (int)xObject.Attribute("x");
-                Y = (int)xObject.Attribute("y");
+                X = (double)xObject.Attribute("x");
+                Y = (double)xObject.Attribute("y");
                 Visible = (bool?)xObject.Attribute("visible") ?? true;
-                Width = (int?)xObject.Attribute("width") ?? 0;
-                Height = (int?)xObject.Attribute("height") ?? 0;
+                Width = (double?)xObject.Attribute("width") ?? 0.0;
+                Height = (double?)xObject.Attribute("height") ?? 0.0;
                 Rotation = (double?)xObject.Attribute("rotation") ?? 0.0;
 
                 // Assess object type and assign appropriate content
@@ -102,18 +102,18 @@ namespace TiledSharp
                 Properties = new PropertyDict(xObject.Element("properties"));
             }
 
-            public List<Tuple<int, int>> ParsePoints(XElement xPoints)
+            public List<Tuple<double, double>> ParsePoints(XElement xPoints)
             {
-                var points = new List<Tuple<int, int>>();
+                var points = new List<Tuple<double, double>>();
 
                 var pointString = (string)xPoints.Attribute("points");
                 var pointStringPair = pointString.Split(' ');
                 foreach (var s in pointStringPair)
                 {
                     var pt = s.Split(',');
-                    var x = int.Parse(pt[0]);
-                    var y = int.Parse(pt[1]);
-                    points.Add(Tuple.Create<int, int>(x, y));
+                    var x = double.Parse(pt[0]);
+                    var y = double.Parse(pt[1]);
+                    points.Add(Tuple.Create<double, double>(x, y));
                 }
                 return points;
             }
